+++ lib/wifi-wand/models/mac_os_model.rb
@@ -162,10 +162,10 @@ class MacOsModel < BaseModel
   def connected_network_name
     return nil unless wifi_on? # no need to try
 
-    command_output = run_os_command("networksetup -getairportnetwork #{wifi_interface}")
-    connected_prefix = 'Current Wi-Fi Network: '
-    connected = Regexp.new(connected_prefix).match?(command_output)
-    connected ? command_output.split(connected_prefix).last.chomp : nil
+    command_output = run_os_command("ipconfig getsummary #{wifi_interface} | grep ' SSID :'", false)
+    return nil if command_output.nil?
+
+    command_output.split('SSID :').last.strip
   end